Extract contextual token splits and vocative addresses
